test(e2e): run on ephemeral clusters
Rework the e2e test to expect a working connection to a cluster when
they start. Developers can create their own clusters and run the tests.

Removed the code used to start kind clusters within the e2e tests.

Reworked the Taskfile to define two environments where the tests can run:
1. An ephemeral one running within Dagger, using the k3s module, to be
used by the CI.
2. A persistent one created with Kind, requiring the kind binary, to be
used for development and debugging when the ephemeral cluster is not
enough.

Signed-off-by: Francesco Canovai <francesco.canovai@enterprisedb.com>