feat: lenient decoding of CNPG resources
This patch allows the barman-cloud plugin to work with operator being
structurally identical with CNPG but with a different API group.

It does that by using lenient decoding of the passed CNPG resources
and by injecting the detected GVK to the sidecar, that uses it to
properly encode and decode the Kubernetes resources.

Signed-off-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>